split project reward adapter
