+++ infrastructure/postgres-adapter/src/main/java/onlydust/com/marketplace/api/postgres/adapter/PostgresProjectRewardV2Adapter.java
@@ -2,36 +2,83 @@ package onlydust.com.marketplace.api.postgres.adapter;
 
 import lombok.AllArgsConstructor;
 import onlydust.com.marketplace.accounting.domain.model.Quote;
+import onlydust.com.marketplace.api.postgres.adapter.entity.read.BudgetStatsEntity;
+import onlydust.com.marketplace.api.postgres.adapter.entity.read.RewardViewEntity;
 import onlydust.com.marketplace.api.postgres.adapter.entity.write.HistoricalQuoteEntity;
-import onlydust.com.marketplace.api.postgres.adapter.repository.CurrencyRepository;
-import onlydust.com.marketplace.api.postgres.adapter.repository.HistoricalQuoteRepository;
-import onlydust.com.marketplace.api.postgres.adapter.repository.ProjectAllowanceRepository;
+import onlydust.com.marketplace.api.postgres.adapter.mapper.RewardMapper;
+import onlydust.com.marketplace.api.postgres.adapter.repository.*;
 import onlydust.com.marketplace.kernel.pagination.Page;
+import onlydust.com.marketplace.kernel.pagination.PaginationHelper;
 import onlydust.com.marketplace.kernel.pagination.SortDirection;
 import onlydust.com.marketplace.project.domain.model.Currency;
 import onlydust.com.marketplace.project.domain.model.Reward;
 import onlydust.com.marketplace.project.domain.port.output.ProjectRewardStoragePort;
 import onlydust.com.marketplace.project.domain.view.*;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 
 import javax.transaction.Transactional;
+import java.math.BigDecimal;
+import java.text.SimpleDateFormat;
 import java.time.Instant;
+import java.util.Collection;
+import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
+import static java.util.Objects.isNull;
 import static onlydust.com.marketplace.kernel.exception.OnlyDustException.internalServerError;
+import static onlydust.com.marketplace.kernel.exception.OnlyDustException.notFound;
 
 @AllArgsConstructor
 public class PostgresProjectRewardV2Adapter implements ProjectRewardStoragePort {
-
     private final ProjectAllowanceRepository projectAllowanceRepository;
     private final HistoricalQuoteRepository historicalQuoteRepository;
     private final CurrencyRepository currencyRepository;
+    private final BudgetStatsRepository budgetStatsRepository;
+    private final RewardViewRepository rewardViewRepository;
+    private final CustomRewardRepository customRewardRepository;
 
     @Override
-    @Transactional
-    public ProjectRewardsPageView findRewards(UUID projectId, ProjectRewardView.Filters filters, Reward.SortBy sortBy, SortDirection sortDirection
-            , int pageIndex, int pageSize) {
-        // TODO
-        return null;
+    @org.springframework.transaction.annotation.Transactional(readOnly = true)
+    public ProjectRewardsPageView findRewards(UUID projectId, ProjectRewardView.Filters filters,
+                                              Reward.SortBy sort, SortDirection sortDirection,
+                                              int pageIndex, int pageSize) {
+        final var format = new SimpleDateFormat("yyyy-MM-dd");
+        final var fromDate = isNull(filters.getFrom()) ? null : format.format(filters.getFrom());
+        final var toDate = isNull(filters.getTo()) ? null : format.format(filters.getTo());
+
+        final var pageRequest = PageRequest.of(pageIndex, pageSize,
+                RewardViewRepository.sortBy(sort, sortDirection == SortDirection.asc ? Sort.Direction.ASC : Sort.Direction.DESC));
+
+        final var page = rewardViewRepository.findProjectRewards(projectId, filters.getCurrencies(), filters.getContributors(), fromDate, toDate, pageRequest);
+        final var budgetStats = budgetStatsRepository.findByProject(projectId, filters.getCurrencies(), filters.getContributors(), fromDate, toDate);
+
+        return ProjectRewardsPageView.builder().
+                rewards(Page.<ProjectRewardView>builder()
+                        .content(page.getContent().stream().map(RewardViewEntity::toProjectReward).toList())
+                        .totalItemNumber((int) page.getTotalElements())
+                        .totalPageNumber(page.getTotalPages())
+                        .build())
+                .remainingBudget(budgetStats.size() == 1 ?
+                        new Money(budgetStats.get(0).getRemainingAmount(),
+                                budgetStats.get(0).getCurrency().toOldDomain(),
+                                budgetStats.get(0).getRemainingUsdAmount()) :
+                        new Money(null, null,
+                                budgetStats.stream().map(BudgetStatsEntity::getRemainingUsdAmount).filter(Objects::nonNull).reduce(BigDecimal.ZERO,
+                                        BigDecimal::add)))
+                .spentAmount(budgetStats.size() == 1 ?
+                        new Money(budgetStats.get(0).getSpentAmount(),
+                                budgetStats.get(0).getCurrency().toOldDomain(),
+                                budgetStats.get(0).getSpentUsdAmount()) :
+                        new Money(null, null,
+                                budgetStats.stream().map(BudgetStatsEntity::getSpentUsdAmount).filter(Objects::nonNull).reduce(BigDecimal.ZERO,
+                                        BigDecimal::add)))
+                .sentRewardsCount(budgetStats.stream().map(BudgetStatsEntity::getRewardIds).flatMap(Collection::stream).collect(Collectors.toUnmodifiableSet()).size())
+                .rewardedContributionsCount(budgetStats.stream().map(BudgetStatsEntity::getRewardItemIds).flatMap(Collection::stream).flatMap(Collection::stream).collect(Collectors.toUnmodifiableSet()).size())
+                .rewardedContributorsCount(budgetStats.stream().map(BudgetStatsEntity::getRewardRecipientIds).flatMap(Collection::stream).collect(Collectors.toUnmodifiableSet()).size())
+                .build();
     }
 
     @Override
@@ -63,17 +110,27 @@ public class PostgresProjectRewardV2Adapter implements ProjectRewardStoragePort 
     }
 
     @Override
-    @Transactional
+    @org.springframework.transaction.annotation.Transactional(readOnly = true)
     public RewardDetailsView getProjectReward(UUID rewardId) {
-        // TODO
-        return null;
+        return rewardViewRepository.find(rewardId)
+                .orElseThrow(() -> notFound("Reward %s not found".formatted(rewardId)))
+                .toDomain();
     }
 
     @Override
-    @Transactional
+    @org.springframework.transaction.annotation.Transactional(readOnly = true)
     public Page<RewardItemView> getProjectRewardItems(UUID rewardId, int pageIndex, int pageSize) {
-        // TODO
-        return null;
+        final Integer count = customRewardRepository.countRewardItemsForRewardId(rewardId);
+        final List<RewardItemView> rewardItemViews =
+                customRewardRepository.findRewardItemsByRewardId(rewardId, pageIndex, pageSize)
+                        .stream()
+                        .map(RewardMapper::itemToDomain)
+                        .toList();
+        return Page.<RewardItemView>builder()
+                .content(rewardItemViews)
+                .totalItemNumber(count)
+                .totalPageNumber(PaginationHelper.calculateTotalNumberOfPage(pageSize, count))
+                .build();
     }
 
     @Override
